+++ apps/rig_api/lib/rig_api/controllers/session_blacklist_controller.ex
@@ -8,6 +8,7 @@ defmodule RigApi.SessionBlacklistController do
 
   """
   use RigApi, :controller
+  use PhoenixSwagger
   require Logger
 
   alias RigAuth.Session
@@ -33,6 +34,19 @@ defmodule RigApi.SessionBlacklistController do
 
   # ---
 
+  # Swagger documentation for endpoint GET /v1/session-blacklist/:session-id
+  swagger_path :check_status do
+    get("/v1/session-blacklist/{sessionId}")
+    summary("Getting the current blacklist status")
+    description("Provides the current status of a blacklisted session given in the parameters")
+
+    parameters do
+      sessionId(:path, :string, "The id of the session", required: true)
+    end
+
+    response(200, "Ok", Schema.ref(:SessionBlacklistStatus))
+  end
+
   @doc "Check blacklist status for a specific session id."
   def check_status(%{method: "GET"} = conn, %{"session_id" => session_id}) do
     json(conn, %{
@@ -42,6 +56,25 @@ defmodule RigApi.SessionBlacklistController do
   end
 
   # ---
+
+  # Swagger documentation for endpoint POST /v1/session-blacklist
+  swagger_path :blacklist_session do
+    post("/v1/session-blacklist")
+    summary("Blacklist a new session")
+    description("Blacklists a new session with data given in the body")
+
+    parameters do
+      sessionBlacklist(
+        :body,
+        Schema.ref(:SessionBlacklistRequest),
+        "The details for blacklisting a session",
+        required: true
+      )
+    end
+
+    response(200, "Ok", Schema.ref(:SessionBlacklistResponse))
+    response(400, "Missing value for 'x'")
+  end
 
   @doc "Plug action to add a session id to the session blacklist."
   def blacklist_session(%{method: "POST"} = conn, _params) do
@@ -66,10 +99,10 @@ defmodule RigApi.SessionBlacklistController do
 
   defp parse(body) do
     {:ok,
-     %{
-       session_id: Map.fetch!(body, "sessionId"),
-       ttl_s: body |> Map.fetch!("validityInSeconds") |> String.to_integer()
-     }}
+      %{
+        session_id: Map.fetch!(body, "sessionId"),
+        ttl_s: body |> Map.fetch!("validityInSeconds") |> String.to_integer()
+      }}
   rescue
     e in KeyError ->
       {:error, "Missing value for '#{e.key}'"}
@@ -77,5 +110,64 @@ defmodule RigApi.SessionBlacklistController do
     e in ArgumentError ->
       # This is likely String.to_integer/1, but we don't know for sure.
       {:error, "Invalid request body: #{inspect(e)}"}
+  end
+ 
+  def swagger_definitions do
+    %{
+      SessionBlacklistRequest:
+        swagger_schema do
+          title("Session Blacklist Request")
+          description("Request for blacklisting a session")
+
+          properties do
+            sessionId(:string, "JWT JTI session Id", required: true)
+
+            validityInSeconds(:string, "Seconds how long a session should be blacklisted",
+              required: true
+            )
+          end
+
+          example(%{
+            sessionId: "SomeSessionID123",
+            validityInSeconds: "60"
+          })
+        end,
+      SessionBlacklistResponse:
+        swagger_schema do
+          title("Session Blacklist Response")
+          description("Response for blacklisting a session")
+
+          properties do
+            sessionId(:string, "JWT JTI session Id", required: true)
+
+            validityInSeconds(:string, "Seconds how long a session should be blacklisted",
+              required: true
+            )
+
+            isBlacklisted(:boolean, "Current status of the blacklisted session", required: true)
+          end
+
+          example(%{
+            sessionId: "SomeSessionID123",
+            validityInSeconds: "60",
+            isBlacklisted: true
+          })
+        end,
+      SessionBlacklistStatus:
+        swagger_schema do
+          title("Session Blacklist Status")
+          description("Status for a blacklisted session")
+
+          properties do
+            sessionId(:string, "JWT JTI session Id", required: true)
+            isBlacklisted(:boolean, "Current status of the blacklisted session", required: true)
+          end
+
+          example(%{
+            sessionId: "SomeSessionID123",
+            isBlacklisted: true
+          })
+        end
+    }
   end
 end